pyproject.toml
[project]
name = "tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)"
]
